+++ lib/db/src/schema/index.ts
@@ -6,3 +6,5 @@ export * from "./lead_logs";
 export * from "./projects";
 export * from "./project_files";
 export * from "./dropdown_options";
+export * from "./parsed_quotations";
+export * from "./parsed_sections";